Move test_monotonic.py to new cudf directory structure
